Extract `classifieds::parse` module
